One more byte loop over the state channel table
writeCommandState walks the table to find the channel a command owns, and
that loop counts in a byte. Past 255 channels it never reaches the slot
count, which is a uint16 now - it would spin rather than stop. Missed
because the widening script matched loops named i, and this one is named c.

Harmless at any table a board can afford today, which is why nothing
caught it.

+++ src/BlaeckSerial.cpp
@@ -1246,7 +1246,7 @@ void BlaeckSerial::writeCommandState(const char *command, unsigned long messageI
     // The channel was declared from this same name at registration, so it exists unless the
     // table was full - in which case there is nothing to publish to and the warning was
     // already given there.
-    for (byte c = 0; c < _stateChannelSlots(); c++)
+    for (uint16_t c = 0; c < _stateChannelSlots(); c++)
     {
       if (!_stateChannels[c].inUse || !_stateChannels[c].ownedByCommand)
         continue;
